Added - Extensions can be provided as any collection instead of just a set on initialisation.   Using a List, e.g., will allow the client to provide a preferred order in which to   load and initialise the extensions. Fixed - Dropping derived columns was sending an  command to   the database which resulted in an error. Derived columns only need to be removed   from the  table. - Do not throw NotfoundException in findColumn in Join as this breaks column search   in SQL Server in complex queries where the referred column might be in an outer query.

+++ src/main/java/ma/vi/esql/parser/define/AlterTable.java
@@ -341,8 +341,10 @@ public class AlterTable extends Define<String> {
             }
           }
 
-          con.createStatement().executeUpdate(
-            "ALTER TABLE " + dbName + " DROP COLUMN \"" + drop.columnName() + '"');
+          if (!column.derived()) {
+            con.createStatement().executeUpdate(
+                "ALTER TABLE " + dbName + " DROP COLUMN \"" + drop.columnName() + '"');
+          }
 
 //          con.createStatement().executeUpdate(
 //            "ALTER TABLE " + dbName + " DROP COLUMN IF EXISTS \"" + drop.columnName() + '"');
